Remove bogus connect(s, NULL, 0) call
The only thing it does is error because of a check added in sockargs() in
uipc_syscalls r1.155. As guenther pointed out, this may have been added
because of a misreading of the last sentence of the first paragraph of the
connect(2) manual.

Instead of erroring, this will keep listening if -k is given and otherwise
it will close the socket and exit with success.

ok guenther jeremy