effectively revert revision 1.7.
The original reason was some corner cases around COPY relocations, which
caused problems for miniperl which directly modified environ and accessed
environ via libc functions.

This causes duplicate symbols for environ/__progname in some (poorly written)
apps, but is allowed on most other architectures.

Since the time this was added other arm architecture changes mean that we
don't need it, so remove it.

debugged with kettenis@ and jca@
tested and OK guenther@