Improve responsiveness in OOM situations & make free target checks coherent.
Remove a change introduced in NetBSD to pageout 4 times as many pages as
required to meet the low water mark of free pages.   With todays' Gbs of
RAMs, it makes the pagedaemon hog the CPU for too long when the amount of
free pages is close to none.

ok sthen@, kettenis@