d2i_ECPrivateKey: split parameter setting into a helper
In order to decode a private key, the group must be known in some way.
Typically, the group is encoded in the EC domain parameters, preferably
as a named curve (this is mandatory in PKIX per RFC 5480).

However, the group could be absent because the domain parameters are
OPTIONAL in the ECPrivateKey SEQUENCE. In that case the code falls
back to the group that may already be set on the EC_KEY. Now there is
no way to tell whether that group is the right one...

In any case. Split this thing out of the body of d2i_ECPrivateKey()
to make that function a bit less of an eyesore.

ok jsing