For AMD SEV automatically load psp(4) firmware during vmd(4) startup.
When opening /dev/psp for the first time, load the SEV firmware
located in /etc/firmware/amdsev/ optained via fw_update(8).  The
relevant file is determined by CPU family and model.
If firmware files are not installed or the correct file could not
be determined, PSP will be initialized using the on-chip firmware.
Alternatively a BIOS upgrade could install a newer version.

from hshoexer@; OK mlarkin@