getgrouplist(3): always return the total number of groups found
In case there is not enough storage to store all the groups found, our
implementation returned the size passed in *ngroups.  This isn't very
useful and is different from other implementations.  Back in 1999 the
NetBSD implementation was changed to return the actual number of groups
found; that lets callers allocate a properly sized array, a behavior was
later copied by FreeBSD and glibc.  So align our implementation. The
documentation doesn't need a change.

ok millert@ gilles@