Neighbor (fault ahead) pages are never mapped with the wired attribute.
Wired faults are always "narrow".  That means the fault handler do not try to
fault neighbor pages ahead.  So do not propagate the `flt->wired' attribute to
the corresponding pmap_enter(9) calls and instead assert that it is false
whenever neighbor pages are entered in a memory space.

ok tb@