sys/uvideo: skip uvideo_vs_set_alt for bulk endpoint
Section 2.4.3 of the UVC 1.5 class specification states that the bulk
endpoint only supports the alternative setting of zero, which is the
default stream, and which is switched at uvideo_attach_hook.

Inside uvideo_vs_close, the code uses the same switch to the alternative
setting of zero to turn off the cam LED.

The additional uvideo_vs_set_alt inside uvideo_vs_open turns off the cam
LED in the case of the bulk endpoint cam. I see this behavior on both
available bulk endpoint cams: it flashes the LED flashes after ffplay
starts, and looks like it was turned off.

OK mglocker@ kn@